+++ Assets/Project/Characters/Humanoid/AI/Pathfinding/PathRequestManager.cs
@@ -65,13 +65,13 @@ public struct PathRequest {
 	public Vector3 pathEnd;
 	public Action<Vector3[], bool> callback;
     public PathfinderStrategy strategy;
-    public int maxLength;
+    public float maxLength;
 
 	public PathRequest(Vector3 _start, 
                        Vector3 _end, 
                        Action<Vector3[], bool> _callback,
                        PathfinderStrategy _strategy,
-                       int _maxLength) {
+                       float _maxLength) {
 		pathStart = _start;
 		pathEnd = _end;
 		callback = _callback;
